UPDATE Books create,update,delete 유저 추가(본인만 수정가능하게 변경) [이성훈]

+++ books/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Book, Book_Review, Book_Review_Comment
 from .forms import Book_ReviewForm, Book_Review_CommentForm
+from django.http import HttpResponseForbidden
+from django.contrib.auth.decorators import login_required
 
 # 책 리스트 페이지
 def index(request):
@@ -35,7 +37,8 @@ def review(request):
     return render(request, "books/review.html", context)
 
 
-# 책 리뷰 작성(user 추가 전)
+# 책 리뷰 작성
+@login_required
 def create(request):
     if request.user.is_authenticated:
         if request.method == "POST":
@@ -59,23 +62,31 @@ def review_detail(request, pk):
     }
     return render(request, "books/review_detail.html", context)
 
-# 리뷰 업데이트 (유저확인 전)
+# 리뷰 업데이트
+@login_required
 def update(request, pk):
     book_review = Book_Review.objects.get(pk=pk)
-    if request.method == "POST":
-        book_review_form = Book_ReviewForm(request.POST, request.FILES, instance=book_review)
-        if book_review_form.is_valid():
-            book_review_form.save()
-            return redirect("books:review_detail", pk)
+    if request.user == book_review.user:
+        if request.method == "POST":
+            book_review_form = Book_ReviewForm(request.POST, request.FILES, instance=book_review)
+            if book_review_form.is_valid():
+                book_review_form.save()
+                return redirect("books:review_detail", pk)
+        else:
+            book_review_form = Book_ReviewForm(instance=book_review)
+        context = {"book_review_form": book_review_form}
+        return render(request, "books/create.html", context)
     else:
-        book_review_form = Book_ReviewForm(instance=book_review)
-    context = {"book_review_form": book_review_form}
-    return render(request, "books/create.html", context)
+        return HttpResponseForbidden()
 
-# 리뷰 삭제 (유저확인 전)
+# 리뷰 삭제
+@login_required
 def delete(request, pk):
     book_review = Book_Review.objects.get(pk=pk)
-    book_review.delete()
-    return redirect("books:review")
+    if request.user == book_review.user:
+        book_review.delete()
+        return redirect("books:review")
+    else:
+        return HttpResponseForbidden()
